The New and Improved C created calculator
The first thing to show up in the function is a calculator menu. The menu has 9  * options to choose from. Enter Values, add, subtract, multiply, divide,   * average, maximum value an option to determine whether each value entered is   * positive, negative, or zero, and a quit option.   * The function presents the menu and requests input from the keyboard.   *	If the user enters an input that is not prompted in the menu then  * an error message will pop up informing the user that they had entered an  * invalid entry. The function has a loop to try again. For a few of the  * cases the user cannot select them without first selecting "Enter Values".  * If the user tries to do so without entering the values then an error message  * will pop up saying that they have not entered any values. The user can choose as  * they please until they select 'Q' for quit. After doing so they will be thanked  * for using the application.